moves insert and delete from __init__.py to util.py
