+++ tuid/util.py
@@ -17,8 +17,8 @@ from mo_logs import Log
 from pyLibrary.sql import quote_set, sql_list
 from mo_threads import Till
 from mo_dots import wrap
-from tuid import insert
 
+TIMEOUT = 10
 HG_URL = URL("https://hg.mozilla.org/")
 
 
@@ -166,6 +166,35 @@ def map_to_array(pairs):
         return tuids
     else:
         return None
+
+
+def wait_until(index, condition):
+    timeout = Till(seconds=TIMEOUT)
+    while not timeout:
+        if condition():
+            break
+        index.refresh()
+
+
+def delete(index, filter):
+    index.delete_record(filter)
+    index.refresh()
+    wait_until(
+        index, lambda: index.search({"size": 0, "query": filter}).hits.total == 0
+    )
+
+
+def insert(index, records):
+    ids = records.value._id
+    index.extend(records)
+    index.refresh()
+    wait_until(
+        index,
+        lambda: index.search(
+            {"size": 0, "query": {"terms": {"_id": ids}}}
+        ).hits.total
+        == len(records),
+    )
 
 
 # Used for increasing readability
